Support shareable JSI installed host functions/objects
This commit adds support for sharing JSI installed host objects and host functions as shared objects.

This is implemented as simply storing the underlying cpp object or function in a value wrapper when adaptet. To convert back a new host object/function is created from the same underlying objects.

It is up to the developer to support accessing values both from multiple threads and from multiple Javascript runtimes.